SixthEdition/requirements.txt
flask==1.1.4
psycopg2-binary==2.8.6
python-dotenv==0.17.1
flask-cors==3.0.10
markupsafe==2.0.1
